# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and profiles gateway",
        "title": "Gandalf API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api/v0",
    "paths": {
...
        "/users/{user_id}/profile": {
            "get": {
                "description": "'This can only be done by the logged in user.'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Get user profile",
                "operationId": "getUserProfileById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/profiles.Profile"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Should be replaced with a PATCH request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Update user profile",
                "operationId": "updateProfileByUserId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update user profile",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile Updated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/profiles.Profile"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "This can only be done by the logged in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Create user profile",
                "operationId": "createProfileByUserId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create user profile",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Profile created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/profiles.Profile"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/{user_id}/profile/full": {
            "get": {
                "description": "'This can only be done by the logged in user.'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Get full user profile",
                "operationId": "getFullProfileByUserId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full profile found\"'",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/profiles.FullProfile"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/{user_id}/profile/recommendation": {
            "get": {
                "description": "'This can only be done by the logged in user.'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Get random profile recommendation",
                "operationId": "getRecommendationByUserId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/profiles.FullProfile"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_ports_http_api_profiles.Prompt": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Spider Man"
                },
                "id": {
                    "type": "string",
                    "example": "75988450-f7c7-4022-b04b-6679e9294056"
                },
                "position": {
                    "type": "integer"
                },
                "question": {
                    "type": "string",
                    "example": "My most irrational fear is..."
                },
                "type": {
                    "type": "string",
                    "example": "text"
                }
            }
        },
        "profiles.FullProfile": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/profiles.Profile"
                },
                "prompts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_ports_http_api_profiles.Prompt"
                    }
                }
            }
        },
        "profiles.Profile": {
            "type": "object",
            "required": [
                "birth_date",
                "drinks_alcohol",
                "family_plans",
                "first_name",
                "has_children",
                "intention",
                "last_name",
                "preferred_partner",
                "sex",
                "smokes"
            ],
            "properties": {
                "birth_date": {
                    "type": "string",
                    "format": "date",
                    "example": "1971-06-28"
                },
                "drinks_alcohol": {
                    "type": "string",
                    "example": "sometimes"
                },
                "family_plans": {
                    "type": "string",
                    "example": "not sure yet"
                },
                "first_name": {
                    "type": "string",
                    "example": "Elon"
                },
                "has_children": {
                    "type": "boolean",
                    "example": false
                },
                "height": {
                    "type": "integer",
                    "example": 180
                },
                "intention": {
                    "type": "string",
                    "example": "long-term relationship"
                },
                "last_name": {
                    "type": "string",
                    "example": "Musk"
                },
                "location": {
                    "type": "string"
                },
                "preferred_partner": {
                    "type": "string",
                    "example": "woman"
                },
                "profile_pic": {
                    "type": "string",
                    "example": "http://example/profile_pic"
                },
                "sex": {
                    "type": "string",
                    "example": "man"
                },
                "smokes": {
                    "type": "string",
                    "example": "no"
                },
                "user_id": {
                    "type": "string",
                    "example": "d2095501-4295-4cb2-b616-94cd2dc5bfb1"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {}
            }
        }
    }
}